refactor: extract ConversationMobilePanelPage from conversation-main
