+++ src/routes/conversation.tsx
@@ -21,10 +21,8 @@ import { useTaskPolling } from "#/hooks/query/use-task-polling";
 
 import { displayErrorToast } from "#/utils/custom-toast-handlers";
 import { useIsAuthed } from "#/hooks/query/use-is-authed";
-import {
-  ConversationMain,
-  ConversationMobilePanelPage,
-} from "#/components/features/conversation/conversation-main/conversation-main";
+import { ConversationMain } from "#/components/features/conversation/conversation-main/conversation-main";
+import { ConversationMobilePanelPage } from "#/components/features/conversation/conversation-main/conversation-mobile-panel-page";
 
 import { WebSocketProviderWrapper } from "#/contexts/websocket-provider-wrapper";
 import { useErrorMessageStore } from "#/stores/error-message-store";
